refactor: to solve block-base problems
